refactor code structure to move cadence worker code to new michelangelo-worker
