+++ go/worker/activities/trigger/BUILD.bazel
@@ -1,0 +1,21 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "module.go",
+        "trigger_activities.go",
+    ],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/activities/trigger",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//go/components/triggerrun:go_default_library",
+        "//go/worker/activities/trigger/parameter:go_default_library",
+        "//proto/api/v2:go_default_library",
+        "@com_github_cadence_workflow_starlark_worker//activity:go_default_library",
+        "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
+        "@com_github_cadence_workflow_starlark_worker//workflow:go_default_library",
+        "@org_uber_go_fx//:go_default_library",
+        "@org_uber_go_zap//:go_default_library",
+    ],
+)
