+++ go/worker/activities/trigger/parameter/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "cron_trigger.go",
+        "parameter_generator.go",
+    ],
+    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/activities/trigger/parameter",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//proto/api/v2:go_default_library",
+    ],
+)
