+++ go/worker/workflows/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//go/worker/workflows/ray:go_default_library",
+        "//go/worker/workflows/trigger:go_default_library",
         "@org_uber_go_fx//:go_default_library",
     ],
 )
