+++ go/worker/workflows/module.go
@@ -2,9 +2,11 @@ package workflows
 
 import (
 	"github.com/michelangelo-ai/michelangelo/go/worker/workflows/ray"
+	"github.com/michelangelo-ai/michelangelo/go/worker/workflows/trigger"
 	"go.uber.org/fx"
 )
 
 var Module = fx.Options(
 	ray.Module,
+	trigger.Module,
 )
