+++ go/worker/workflows/trigger/BUILD.bazel
@@ -3,20 +3,22 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
+        "cron_trigger_workflows.go",
         "module.go",
-        "trigger.go",
     ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/components/triggerrun/workflow",
+    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/workflows/trigger",
     visibility = ["//visibility:public"],
     deps = [
+        "//go/components/triggerrun:go_default_library",
         "//go/components/utils:go_default_library",
+        "//go/worker/activities/trigger:go_default_library",
+        "//go/worker/activities/trigger/parameter:go_default_library",
         "//proto/api:go_default_library",
         "//proto/api/v2:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
         "@com_github_gogo_protobuf//types:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/types:go_default_library",
-        "@org_uber_go_cadence//activity:go_default_library",
         "@org_uber_go_cadence//workflow:go_default_library",
         "@org_uber_go_fx//:go_default_library",
         "@org_uber_go_zap//:go_default_library",
