+++ python/michelangelo/cli/sandbox/demo/triggerrun.yaml
@@ -1,0 +1,18 @@
+apiVersion: michelangelo.api/v2
+kind: TriggerRun
+metadata:
+  name: training-pipeline-cron-trigger
+  namespace: ma-dev-test
+  annotations:
+    description: "Test training pipeline execution via cron trigger"
+spec:
+  pipeline:
+    name: training-pipeline
+    namespace: ma-dev-test
+  trigger:
+    cronSchedule:
+      cron: "* * * * *"  # Run every minute for testing
+    # batchPolicy:
+    #   batchSize: 1
+    #   wait: "60s"
+    maxConcurrency: 1
